Refactor: Pass mutation to coy route and messages via context
