+++ react/components/EditorContainer/Topbar/ContextSelectors/BindingCloning/withBindingsContext.tsx
@@ -16,7 +16,7 @@ import {
 
 import GetRouteQuery from '../graphql/GetRoute.graphql'
 import {
-  BindingSelectorItem,
+  // BindingSelectorItem,
   BindingSelectorState,
   setInitialBindingState,
   useBindingSelectorReducer,
@@ -47,12 +47,15 @@ interface CloneContentContextValue {
     routeInfo?: Route
     bindingSelector: BindingSelectorState
     currentBinding: Binding
+    pageContext: PageContext
   }
   actions: {
     refetchRouteInfo: (
       variables?: OperationVariables
     ) => Promise<ApolloQueryResult<any>>
     dispatchBindingSelector: Dispatch<BindingSelectorAction>
+    saveRoute: (args: MutationArgs<SaveRouteVariables>) => void
+    copyBindings: (args: MutationArgs<CopyBindingVariables>) => void
   }
 }
 
@@ -60,14 +63,32 @@ const CloneContentContext = createContext<CloneContentContextValue>(
   {} as CloneContentContextValue
 )
 
+interface SaveRouteVariables {
+  route: Route
+}
+
+type MutationArgs<T> = {
+  variables: T
+}
+
+interface CopyBindingVariables {
+  from: string
+  to: string
+  template: string
+  context: PageContext
+}
+
 interface WithBindingsQueriesProps {
   routeInfoQuery: QueryResult & { route?: Route }
+  saveRoute: (args: MutationArgs<SaveRouteVariables>) => void
+  copyBindings: (args: MutationArgs<CopyBindingVariables>) => void
 }
 
 interface CloneContentProps {
   routeId: string
   bindings: Binding[]
   currentBinding: Binding
+  pageContext: PageContext
 }
 
 const CloneContentProvider: FC<CloneContentProps &
@@ -76,9 +97,11 @@ const CloneContentProvider: FC<CloneContentProps &
   routeInfoQuery,
   bindings,
   currentBinding,
+  pageContext,
+  saveRoute,
+  copyBindings,
 }) => {
   const [bindingSelector, dispatchBindingSelector] = useBindingSelectorReducer()
-
   const {
     route: routeInfo,
     loading,
@@ -98,8 +121,20 @@ const CloneContentProvider: FC<CloneContentProps &
   return (
     <CloneContentContext.Provider
       value={{
-        data: { loading, error, routeInfo, bindingSelector, currentBinding },
-        actions: { refetchRouteInfo, dispatchBindingSelector },
+        data: {
+          loading,
+          error,
+          routeInfo,
+          bindingSelector,
+          currentBinding,
+          pageContext,
+        },
+        actions: {
+          refetchRouteInfo,
+          dispatchBindingSelector,
+          saveRoute,
+          copyBindings,
+        },
       }}
     >
       {children}
@@ -142,9 +177,9 @@ const BindingFormatter = ({
 interface BindingsContext {
   saveRoute: () => any
   copyBindings: () => any
-  dispatch: () => any
-  state: BindingSelectorItem[]
-  pageContext: PageContext
+  // dispatch: () => any
+  // state: BindingSelectorItem[]
+  // pageContext: PageContext
 }
 
 // TODO: use the proper React context API
@@ -153,15 +188,15 @@ interface BindingsContext {
 // which caused the modal to blink.
 const BindingsContext = withBindingsQueries(
   ({
-    bindings,
+    // bindings,
     routeInfoQuery,
-    currentBinding,
+    // currentBinding,
     children,
     ...props
   }: {
-    bindings: Binding[]
+    // bindings: Binding[]
     routeInfoQuery: any
-    currentBinding: Binding
+    // currentBinding: Binding
     children: (...args: any) => ReactNode
   }) => {
     const { route } = routeInfoQuery
@@ -180,28 +215,23 @@ const BindingsContext = withBindingsQueries(
 
 // re. the dangling comma after T https://stackoverflow.com/a/56989122
 const withBindingsData = <T,>(Component: FunctionComponent<T>) => ({
-  bindings,
-  iframeRuntime,
-  currentBinding,
+  // bindings,
+  // iframeRuntime,
+  // currentBinding,
   ...props
 }: {
-  bindings: Binding[]
-  currentBinding: Binding
-  iframeRuntime: RenderContext
+  // bindings: Binding[]
+  // currentBinding: Binding
+  // iframeRuntime: RenderContext
 } & Omit<T, keyof BindingsContext>) => {
-  const route = iframeRuntime?.route
-  const { id: routeId, pageContext } = route ?? {}
-
   return (
     <BindingsContext
-      currentBinding={currentBinding}
-      bindings={bindings}
-      routeId={routeId}
-      key={routeId}
+    // currentBinding={currentBinding}
+    // bindings={bindings}
+    // routeId={routeId}
+    // key={routeId}
     >
-      {(result: any) => (
-        <Component {...{ currentBinding, pageContext, ...result, ...props }} />
-      )}
+      {(result: any) => <Component {...{ ...result, ...props }} />}
     </BindingsContext>
   )
 }
